optimize the code of Home_widgets screen
