+++ lib/Screen/Home_widgets/catalogHeader.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+import 'package:myfirstproject/widgets/theme.dart';
+import 'package:velocity_x/velocity_x.dart';
+
+class CatalogHeader extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Column(crossAxisAlignment: CrossAxisAlignment.start, children: [
+      "Catalog App".text.xl5.bold.color(MyTheme.darkBluisColor).make(),
+      "Trending Products".text.xl2.make()
+    ]);
+  }
+}
